Top menu new style plus new filter component
Move the sidebar to a top menu
Animate submenus as droping cards
Submenus appear on hover (this can be changed)

Slidoshow is slightly modified:
-- Styled left and right arrows with hover effect
-- slight fade to black on screen edges
-- height slightly reduced

Filter by price: new component
-- It seats below the slideshow
-- more compact look and feel

TODO: Improve the actions for filtering and ordering and connect to filter component
TODO: Close submenus when the mouse goes up into the navbar
TODO: Check on different screen sizes, and set a max-width for the slideshow maybe.
TODO: Maybe the slideshow should be thinner.